Consolidate samples
