+++ sample/pi_2_dht_read.c
